🚚 Move Separator component
