+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Elements/index.js
@@ -10,3 +10,4 @@ export { default as Header } from './Header';
 export { default as SupplementaryLabel } from './SupplementaryLabel';
 export { default as Title } from './Title';
 export { default as TitleWrapper } from './TitleWrapper';
+export { default as Separator } from './Separator';
